media: hantro: Add H.264 encoding support
This is a forward-port of changes at [1] from v5.11 to v6.0 with no
substantial functional changes. I've fixed up small portions of the
patch to account for changes in the original sources, and have resolved
some checkpatch findings.

In full transparency, I have little to no understanding of how this
implementation actually works, and no access to any detailed reference
for the Hantro H1 hardware that might give me a good starting point to
properly explain it. However, I have tested the driver on Rockchip
RK3399-based hardware (Pine64 ROCKPro64 v2.1) with [2] and have found
that it appears to behave as expected.

[1] https://github.com/bootlin/linux
[2] https://github.com/bootlin/v4l2-hantro-h264-encoder

Co-developed-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Signed-off-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Signed-off-by: Alex Hamlin <alex@alexhamlin.co>